demo with csv offline db

+++ utils/utils.py
@@ -8,7 +8,7 @@ def set_logging():
     logger.setLevel(logging.INFO)
 
 
-def make_dirs(dir_root):
-    if os.path.exists(dir_root):
+def make_dirs(dir_root, reset=False):
+    if os.path.exists(dir_root) and reset:
         shutil.rmtree(dir_root)
     os.makedirs(os.path.join(dir_root), exist_ok=True)
